feat: Add variable penalty proportional to distance
This is trying to make the penalty a bit smarter.

The distance is manhattan distance.

The issue with this implementation is clearer now, the penalties are only
applied to the *initial* best tile images.  That means that we still end up with
large areas of duplicate tiles, just not the original best tile.